code cleanup & enable trace ("-t") for MBR decoding ("-mbr")
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk@3246 1f5c12ca-751b-0410-a591-d2e778427230

+++ moses/src/BitmapContainer.cpp
@@ -147,7 +147,7 @@ BackwardsEdge::BackwardsEdge(const BitmapContainer &prevBitmapContainer
 		// initial position is not further into the sentence than the distortion limit.
 		if (hypo.GetWordsBitmap().GetNumWordsCovered() == 0)
 			{
-				if (transOptRange.GetStartPos() <= maxDistortion)
+				if ((int)transOptRange.GetStartPos() <= maxDistortion)
 					m_hypotheses.push_back(&hypo);
 			}
 		else
